refactor: example projects
