+++ Examples/JoinHiderBot/JoinHiderBot.csproj
@@ -3,10 +3,12 @@
     <PropertyGroup>
         <OutputType>Exe</OutputType>
         <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>disable</Nullable>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="TelegramBotBase" Version="5.0.0"/>
+        <ProjectReference Include="..\..\TelegramBotBase\TelegramBotBase.csproj"/>
     </ItemGroup>
 
 </Project>
