+++ Examples/SystemCommandsBot/SystemCommandsBot.csproj
@@ -3,11 +3,12 @@
     <PropertyGroup>
         <OutputType>Exe</OutputType>
         <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>disable</Nullable>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Newtonsoft.Json" Version="13.0.1"/>
-        <PackageReference Include="TelegramBotBase" Version="5.0.0"/>
+        <ProjectReference Include="..\..\TelegramBotBase\TelegramBotBase.csproj"/>
     </ItemGroup>
 
 </Project>
