+++ TelegramBotBase.Test/Tests/Controls/ButtonGridForm.cs
@@ -4,7 +4,7 @@ using TelegramBotBase.Controls.Hybrid;
 using TelegramBotBase.Enums;
 using TelegramBotBase.Form;
 
-namespace TelegramBotBaseTest.Tests.Controls;
+namespace TelegramBotBase.Example.Tests.Controls;
 
 public class ButtonGridForm : AutoCleanForm
 {
