+++ TelegramBotBase.Test/Tests/Controls/ButtonGridPagingForm.cs
@@ -6,7 +6,7 @@ using TelegramBotBase.Controls.Hybrid;
 using TelegramBotBase.Enums;
 using TelegramBotBase.Form;
 
-namespace TelegramBotBaseTest.Tests.Controls;
+namespace TelegramBotBase.Example.Tests.Controls;
 
 public class ButtonGridPagingForm : AutoCleanForm
 {
