+++ TelegramBotBase.Test/Tests/Controls/ButtonGridTagForm.cs
@@ -8,7 +8,7 @@ using TelegramBotBase.DataSources;
 using TelegramBotBase.Enums;
 using TelegramBotBase.Form;
 
-namespace TelegramBotBaseTest.Tests.Controls;
+namespace TelegramBotBase.Example.Tests.Controls;
 
 public class ButtonGridTagForm : AutoCleanForm
 {
