+++ TelegramBotBase.Test/Tests/Controls/CheckedButtonListForm.cs
@@ -6,7 +6,7 @@ using TelegramBotBase.DataSources;
 using TelegramBotBase.Enums;
 using TelegramBotBase.Form;
 
-namespace TelegramBotBaseTest.Tests.Controls;
+namespace TelegramBotBase.Example.Tests.Controls;
 
 public class CheckedButtonListForm : AutoCleanForm
 {
