+++ TelegramBotBase.Test/Tests/Controls/ToggleButtons.cs
@@ -5,7 +5,7 @@ using TelegramBotBase.Controls.Inline;
 using TelegramBotBase.Enums;
 using TelegramBotBase.Form;
 
-namespace TelegramBotBaseTest.Tests.Controls;
+namespace TelegramBotBase.Example.Tests.Controls;
 
 public class ToggleButtons : AutoCleanForm
 {
