+++ TelegramBotBase.Test/Tests/Groups/WelcomeUser.cs
@@ -6,7 +6,7 @@ using TelegramBotBase.Args;
 using TelegramBotBase.Base;
 using TelegramBotBase.Form;
 
-namespace TelegramBotBaseTest.Tests.Groups;
+namespace TelegramBotBase.Example.Tests.Groups;
 
 public class WelcomeUser : GroupForm
 {
