+++ TelegramBotBase.Test/Tests/Navigation/Form1.cs
@@ -4,7 +4,7 @@ using Telegram.Bot.Types;
 using TelegramBotBase.Base;
 using TelegramBotBase.Form;
 
-namespace TelegramBotBaseTest.Tests.Navigation;
+namespace TelegramBotBase.Example.Tests.Navigation;
 
 public class Form1 : FormBase
 {
