+++ TelegramBotBase.Test/Tests/Register/Steps/Step1.cs
@@ -3,7 +3,7 @@ using TelegramBotBase.Args;
 using TelegramBotBase.Base;
 using TelegramBotBase.Form;
 
-namespace TelegramBotBaseTest.Tests.Register.Steps;
+namespace TelegramBotBase.Example.Tests.Register.Steps;
 
 public class Step1 : AutoCleanForm
 {
