+++ TelegramBotBase.Test/Tests/SimpleForm.cs
@@ -4,7 +4,7 @@ using TelegramBotBase.Base;
 using TelegramBotBase.Enums;
 using TelegramBotBase.Form;
 
-namespace TelegramBotBaseTest.Tests;
+namespace TelegramBotBase.Example.Tests;
 
 public class SimpleForm : AutoCleanForm
 {
