+++ TelegramBotBase.Test/Tests/TestForm.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using TelegramBotBase.Base;
 using TelegramBotBase.Form;
 
-namespace TelegramBotBaseTest.Tests;
+namespace TelegramBotBase.Example.Tests;
 
 public class TestForm : FormBase
 {
