--- TelegramBotBase.Test/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-    <package id="Newtonsoft.Json" version="12.0.2" targetFramework="net5"/>
-    <package id="System.Net.Requests" version="4.3.0" targetFramework="net5"/>
-    <package id="Telegram.Bot" version="14.12.0" targetFramework="net5"/>
-</packages>
